A bit of refactoring, testing
